+++ e3nn/math/soft_heaviside.py
@@ -1,0 +1,56 @@
+import torch
+
+
+class _SoftUnitStep(torch.autograd.Function):
+    @staticmethod
+    def forward(ctx, x):
+        ctx.save_for_backward(x)
+        y = torch.zeros_like(x)
+        m = (x > 0.0)
+        y[m] = (-1/x[m]).exp()
+        return y
+
+    @staticmethod
+    def backward(ctx, dy):
+        x, = ctx.saved_tensors
+        dx = torch.zeros_like(x)
+        m = (x > 0.0)
+        xm = x[m]
+        dx[m] = (-1/xm).exp() / xm.pow(2)
+        return dx * dy
+
+
+def soft_unit_step(x):
+    r"""smooth :math:`C^\infty` version of the unit step function
+
+    .. math::
+
+        x \mapsto \theta(x) e^{-1/x}
+
+
+    Parameters
+    ----------
+    x : `torch.Tensor`
+        tensor of shape :math:`(...)`
+
+    Returns
+    -------
+    `torch.Tensor`
+        tensor of shape :math:`(...)`
+
+    Examples
+    --------
+
+    .. jupyter-execute::
+        :hide-code:
+
+        import torch
+        from e3nn.math import soft_unit_step
+        import matplotlib.pyplot as plt
+
+    .. jupyter-execute::
+
+        x = torch.linspace(-1.0, 10.0, 1000)
+        plt.plot(x, soft_unit_step(x));
+    """
+    return _SoftUnitStep.apply(x)