+++ ChangeLog.md
@@ -6,12 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- `math.soft_unit_step` function
 - `nn.model.v2103` generic message passing model + examples of networks using it.
 - `o3.TensorProduct`: is jit scriptable
 - `o3.TensorProduct`: also broadcast the `weight` argument
 - simple e3nn models can be saved/loaded with `torch.save()`/`torch.load()`
 - JITable `o3.SphericalHarmonics` module version of `o3.spherical_harmonics`
 - `in_place` option for `e3nn.util.jit` compilation functions
+- New `@compile_mode("unsupported")` for modules that do not support TorchScript
 
 ### Changed
 - e3nn now requires the latest stable PyTorch, >=1.8.0